+++ SpatialGDK/Source/SpatialGDK/Private/Utils/SpatialDebugger.cpp
@@ -33,9 +33,15 @@ DEFINE_LOG_CATEGORY(LogSpatialDebugger);
 
 namespace
 {
+// Background material for worker region
 const FString DEFAULT_WORKER_REGION_MATERIAL =
 	TEXT("/SpatialGDK/SpatialDebugger/Materials/TranslucentWorkerRegion.TranslucentWorkerRegion");
-}
+// Improbable primary font - Muli regular
+const FString DEFAULT_WORKER_TEXT_FONT = TEXT("/SpatialGDK/SpatialDebugger/Fonts/MuliFont.MuliFont");
+// Material to combine both the background and the worker information in one material
+const FString DEFAULT_WORKER_COMBINED_MATERIAL =
+	TEXT("/SpatialGDK/SpatialDebugger/Materials/WorkerRegionCombinedMaterial.WorkerRegionCombinedMaterial");
+} // anonymous namespace
 
 ASpatialDebugger::ASpatialDebugger(const FObjectInitializer& ObjectInitializer)
 	: Super(ObjectInitializer)
@@ -51,6 +57,8 @@ ASpatialDebugger::ASpatialDebugger(const FObjectInitializer& ObjectInitializer)
 	NetUpdateFrequency = 1.f;
 
 	NetDriver = Cast<USpatialNetDriver>(GetNetDriver());
+
+	OnConfigUIClosed.BindDynamic(this, &ASpatialDebugger::DefaultOnConfigUIClosed);
 
 	// For GDK design reasons, this is the approach chosen to get a pointer
 	// on the net driver to the client ASpatialDebugger.  Various alternatives
@@ -164,10 +172,12 @@ void ASpatialDebugger::OnAuthorityGained()
 			WorkerRegions.SetNum(LBStrategyRegions.Num());
 			for (int i = 0; i < LBStrategyRegions.Num(); i++)
 			{
+				FWorkerRegionInfo WorkerRegionInfo;
 				const TPair<VirtualWorkerId, FBox2D>& LBStrategyRegion = LBStrategyRegions[i];
+				WorkerRegionInfo.VirtualWorkerID = LBStrategyRegion.Key;
 				const PhysicalWorkerName* WorkerName =
 					NetDriver->VirtualWorkerTranslator->GetPhysicalWorkerForVirtualWorker(LBStrategyRegion.Key);
-				FWorkerRegionInfo WorkerRegionInfo;
+				WorkerRegionInfo.WorkerName = (WorkerName == nullptr) ? "" : *WorkerName;
 				WorkerRegionInfo.Color = (WorkerName == nullptr) ? InvalidServerTintColor : SpatialGDK::GetColorForWorkerName(*WorkerName);
 				WorkerRegionInfo.Extents = LBStrategyRegion.Value;
 				WorkerRegions[i] = WorkerRegionInfo;
@@ -186,17 +196,41 @@ void ASpatialDebugger::CreateWorkerRegions()
 		return;
 	}
 
+	UMaterial* WorkerCombinedMaterial = LoadObject<UMaterial>(nullptr, *DEFAULT_WORKER_COMBINED_MATERIAL);
+	if (WorkerCombinedMaterial == nullptr)
+	{
+		UE_LOG(LogSpatialDebugger, Error, TEXT("Worker regions were not rendered. Could not find default material: %s"),
+			   *DEFAULT_WORKER_COMBINED_MATERIAL);
+	}
+
+	UFont* WorkerInfoFont = LoadObject<UFont>(nullptr, *DEFAULT_WORKER_TEXT_FONT);
+	if (WorkerInfoFont == nullptr)
+	{
+		UE_LOG(LogSpatialDebugger, Error, TEXT("Worker information was not rendered. Could not find default font: %s"),
+			   *DEFAULT_WORKER_TEXT_FONT);
+	}
+
 	// Create new actors for all new worker regions
 	FActorSpawnParameters SpawnParams;
 	SpawnParams.bNoFail = true;
+	UWorld* World = GetWorld();
 #if WITH_EDITOR
+	if (World == nullptr)
+	{
+		// We are in the editor at design time
+		World = GEditor->GetEditorWorldContext().World();
+	}
 	SpawnParams.bHideFromSceneOutliner = true;
 #endif
+	check(World != nullptr);
 	SpawnParams.SpawnCollisionHandlingOverride = ESpawnActorCollisionHandlingMethod::AlwaysSpawn;
 	for (const FWorkerRegionInfo& WorkerRegionData : WorkerRegions)
 	{
-		AWorkerRegion* WorkerRegion = GetWorld()->SpawnActor<AWorkerRegion>(SpawnParams);
-		WorkerRegion->Init(WorkerRegionMaterial, WorkerRegionData.Color, WorkerRegionData.Extents, WorkerRegionVerticalScale);
+		AWorkerRegion* WorkerRegion = World->SpawnActor<AWorkerRegion>(SpawnParams);
+		FString WorkerInfo = FString::Printf(TEXT("You are looking at virtual worker number %d\n%s"), WorkerRegionData.VirtualWorkerID,
+											 *WorkerRegionData.WorkerName);
+		WorkerRegion->Init(WorkerRegionMaterial, WorkerCombinedMaterial, WorkerInfoFont, WorkerRegionData.Color, WorkerRegionOpacity,
+						   WorkerRegionData.Extents, WorkerRegionHeight, WorkerRegionVerticalScale, WorkerInfo);
 		WorkerRegion->SetActorEnableCollision(false);
 	}
 }
@@ -282,7 +316,87 @@ void ASpatialDebugger::OnEntityAdded(const Worker_EntityId EntityId)
 		if (Actor->IsA<APlayerController>())
 		{
 			LocalPlayerController = Cast<APlayerController>(Actor);
+
+			if (GetNetMode() == NM_Client)
+			{
+				LocalPlayerController->InputComponent->BindKey(ConfigUIToggleKey, IE_Pressed, this, &ASpatialDebugger::OnToggleConfigUI);
+			}
 		}
+	}
+}
+
+void ASpatialDebugger::OnToggleConfigUI()
+{
+	if (ConfigUIWidget == nullptr)
+	{
+		if (ConfigUIClass != nullptr)
+		{
+			ConfigUIWidget = CreateWidget<USpatialDebuggerConfigUI>(LocalPlayerController.Get(), ConfigUIClass);
+			if (ConfigUIWidget == nullptr)
+			{
+				UE_LOG(LogSpatialDebugger, Error,
+					   TEXT("SpatialDebugger config UI will not load. Couldn't create config UI widget for class: %s"),
+					   *GetNameSafe(ConfigUIClass));
+				return;
+			}
+			else
+			{
+				ConfigUIWidget->SetSpatialDebugger(this);
+			}
+		}
+		else
+		{
+			UE_LOG(LogSpatialDebugger, Error,
+				   TEXT("SpatialDebugger config UI will not load. ConfigUIClass is not set on the spatial debugger."));
+			return;
+		}
+
+		ConfigUIWidget->AddToViewport();
+
+		FInputModeGameAndUI InputModeSettings;
+		InputModeSettings.SetLockMouseToViewportBehavior(EMouseLockMode::DoNotLock);
+		InputModeSettings.SetWidgetToFocus(ConfigUIWidget->TakeWidget());
+
+		LocalPlayerController->SetInputMode(InputModeSettings);
+		LocalPlayerController->bShowMouseCursor = true;
+
+		ConfigUIWidget->OnShow();
+	}
+	else
+	{
+		ConfigUIWidget->RemoveFromParent();
+		ConfigUIWidget = nullptr;
+		OnConfigUIClosed.ExecuteIfBound();
+	}
+}
+
+void ASpatialDebugger::DefaultOnConfigUIClosed()
+{
+	if (LocalPlayerController.IsValid())
+	{
+		FInputModeGameOnly InputModeSettings;
+		LocalPlayerController->SetInputMode(InputModeSettings);
+		LocalPlayerController->bShowMouseCursor = false;
+	}
+}
+
+void ASpatialDebugger::SetShowWorkerRegions(const bool bNewShow)
+{
+	if (bNewShow != bShowWorkerRegions)
+	{
+		if (IsEnabled())
+		{
+			if (bNewShow)
+			{
+				CreateWorkerRegions();
+			}
+			else
+			{
+				DestroyWorkerRegions();
+			}
+		}
+
+		bShowWorkerRegions = bNewShow;
 	}
 }
 
@@ -337,13 +451,10 @@ void ASpatialDebugger::ActorAuthorityIntentChanged(Worker_EntityId EntityId, Vir
 	NetDriver->Connection->SendComponentUpdate(EntityId, &DebuggingUpdate);
 }
 
-void ASpatialDebugger::DrawTag(UCanvas* Canvas, const FVector2D& ScreenLocation, const Worker_EntityId EntityId, const FString& ActorName)
+void ASpatialDebugger::DrawTag(UCanvas* Canvas, const FVector2D& ScreenLocation, const Worker_EntityId EntityId, const FString& ActorName,
+							   const bool bCentre)
 {
 	SCOPE_CYCLE_COUNTER(STAT_DrawTag);
-
-	// TODO: Smarter positioning of elements so they're centered no matter how many are enabled
-	// https://improbableio.atlassian.net/browse/UNR-2360.
-	int32 HorizontalOffset = -32.0f;
 
 	check(NetDriver != nullptr && !NetDriver->IsServer());
 	if (!NetDriver->StaticComponentView->HasComponent(EntityId, SpatialConstants::SPATIAL_DEBUGGING_COMPONENT_ID))
@@ -353,13 +464,58 @@ void ASpatialDebugger::DrawTag(UCanvas* Canvas, const FVector2D& ScreenLocation,
 
 	const SpatialDebugging* DebuggingInfo = NetDriver->StaticComponentView->GetComponentData<SpatialDebugging>(EntityId);
 
-	static const float BaseHorizontalOffset(16.0f);
-
 	if (!FApp::CanEverRender()) // DrawIcon can attempt to use the underlying texture resource even when using nullrhi
 	{
 		return;
 	}
 
+	static const float BaseHorizontalOffset = 16.0f;
+	static const float NumberScale = 0.75f;
+	static const float TextScale = 0.5f;
+	const float AuthIdWidth = NumberScale * GetNumberOfDigitsIn(DebuggingInfo->AuthoritativeVirtualWorkerId);
+	const float AuthIntentIdWidth = NumberScale * GetNumberOfDigitsIn(DebuggingInfo->IntentVirtualWorkerId);
+	const float EntityIdWidth = NumberScale * GetNumberOfDigitsIn(EntityId);
+
+	int32 HorizontalOffset = 0;
+	if (bCentre)
+	{
+		// If tag should be centered, calculate the total width of the icons and text to be rendered
+		float TagWidth = 0;
+		if (bShowLock)
+		{
+			// If showing the lock, add the lock icon width
+			TagWidth += BaseHorizontalOffset;
+		}
+		if (bShowAuth)
+		{
+			// If showing the authority, add the authority icon width and the width of the authoritative virtual worker ID
+			TagWidth += BaseHorizontalOffset;
+			TagWidth += (BaseHorizontalOffset * AuthIdWidth);
+		}
+		if (bShowAuthIntent)
+		{
+			// If showing the authority intent, add the authority intent icon width and the width of the authoritative intent virtual worker
+			// ID
+			TagWidth += BaseHorizontalOffset;
+			TagWidth += (BaseHorizontalOffset * AuthIntentIdWidth);
+		}
+		if (bShowEntityId)
+		{
+			// If showing the entity ID, add the width of the entity ID
+			TagWidth += (BaseHorizontalOffset * EntityIdWidth);
+		}
+		if (bShowActorName)
+		{
+			// If showing the actor name, add the width of the actor name
+			const float ActorNameWidth = TextScale * ActorName.Len();
+			TagWidth += (BaseHorizontalOffset * ActorNameWidth);
+		}
+
+		// Calculate the offset based on the total width of the tag
+		HorizontalOffset = TagWidth / -2;
+	}
+
+	// Draw icons and text based on the offset
 	if (bShowLock)
 	{
 		SCOPE_CYCLE_COUNTER(STAT_DrawIcons);
@@ -379,13 +535,11 @@ void ASpatialDebugger::DrawTag(UCanvas* Canvas, const FVector2D& ScreenLocation,
 		Canvas->DrawIcon(Icons[ICON_AUTH], ScreenLocation.X + HorizontalOffset, ScreenLocation.Y, 1.0f);
 		HorizontalOffset += BaseHorizontalOffset;
 		Canvas->SetDrawColor(ServerWorkerColor);
-		const float BoxScaleBasedOnNumberSize = 0.75f * GetNumberOfDigitsIn(DebuggingInfo->AuthoritativeVirtualWorkerId);
-		Canvas->DrawScaledIcon(Icons[ICON_BOX], ScreenLocation.X + HorizontalOffset, ScreenLocation.Y,
-							   FVector(BoxScaleBasedOnNumberSize, 1.f, 1.f));
+		Canvas->DrawScaledIcon(Icons[ICON_BOX], ScreenLocation.X + HorizontalOffset, ScreenLocation.Y, FVector(AuthIdWidth, 1.f, 1.f));
 		Canvas->SetDrawColor(GetTextColorForBackgroundColor(ServerWorkerColor));
 		Canvas->DrawText(RenderFont, FString::FromInt(DebuggingInfo->AuthoritativeVirtualWorkerId), ScreenLocation.X + HorizontalOffset + 1,
 						 ScreenLocation.Y, 1.1f, 1.1f, FontRenderInfo);
-		HorizontalOffset += (BaseHorizontalOffset * BoxScaleBasedOnNumberSize);
+		HorizontalOffset += (BaseHorizontalOffset * AuthIdWidth);
 	}
 
 	if (bShowAuthIntent)
@@ -394,15 +548,14 @@ void ASpatialDebugger::DrawTag(UCanvas* Canvas, const FVector2D& ScreenLocation,
 		const FColor& VirtualWorkerColor = DebuggingInfo->IntentColor;
 		Canvas->SetDrawColor(FColor::White);
 		Canvas->DrawIcon(Icons[ICON_AUTH_INTENT], ScreenLocation.X + HorizontalOffset, ScreenLocation.Y, 1.0f);
-		HorizontalOffset += 16.0f;
+		HorizontalOffset += BaseHorizontalOffset;
 		Canvas->SetDrawColor(VirtualWorkerColor);
-		const float BoxScaleBasedOnNumberSize = 0.75f * GetNumberOfDigitsIn(DebuggingInfo->IntentVirtualWorkerId);
 		Canvas->DrawScaledIcon(Icons[ICON_BOX], ScreenLocation.X + HorizontalOffset, ScreenLocation.Y,
-							   FVector(BoxScaleBasedOnNumberSize, 1.f, 1.f));
+							   FVector(AuthIntentIdWidth, 1.f, 1.f));
 		Canvas->SetDrawColor(GetTextColorForBackgroundColor(VirtualWorkerColor));
 		Canvas->DrawText(RenderFont, FString::FromInt(DebuggingInfo->IntentVirtualWorkerId), ScreenLocation.X + HorizontalOffset + 1,
 						 ScreenLocation.Y, 1.1f, 1.1f, FontRenderInfo);
-		HorizontalOffset += (BaseHorizontalOffset * BoxScaleBasedOnNumberSize);
+		HorizontalOffset += (BaseHorizontalOffset * AuthIntentIdWidth);
 	}
 
 	FString Label;
@@ -453,49 +606,67 @@ void ASpatialDebugger::DrawDebug(UCanvas* Canvas, APlayerController* /* Controll
 	}
 #endif
 
-	DrawDebugLocalPlayer(Canvas);
+	if (ActorTagDrawMode >= EActorTagDrawMode::LocalPlayer)
+	{
+		DrawDebugLocalPlayer(Canvas);
+	}
 
 	FVector PlayerLocation = FVector::ZeroVector;
 
-	if (LocalPawn.IsValid())
+	if (ActorTagDrawMode == EActorTagDrawMode::All)
 	{
-		PlayerLocation = LocalPawn->GetActorLocation();
-	}
-
-	for (TPair<Worker_EntityId_Key, TWeakObjectPtr<AActor>>& EntityActorPair : EntityActorMapping)
-	{
-		const TWeakObjectPtr<AActor> Actor = EntityActorPair.Value;
-		const Worker_EntityId EntityId = EntityActorPair.Key;
-
-		if (Actor != nullptr)
+		if (LocalPawn.IsValid())
 		{
-			FVector ActorLocation = Actor->GetActorLocation();
-
-			if (ActorLocation.IsZero())
-			{
-				continue;
-			}
-
-			if (FVector::Dist(PlayerLocation, ActorLocation) > MaxRange)
-			{
-				continue;
-			}
-
-			FVector2D ScreenLocation = FVector2D::ZeroVector;
-			if (LocalPlayerController.IsValid())
-			{
-				SCOPE_CYCLE_COUNTER(STAT_Projection);
-				UGameplayStatics::ProjectWorldToScreen(LocalPlayerController.Get(), ActorLocation + WorldSpaceActorTagOffset,
-													   ScreenLocation, false);
-			}
-
-			if (ScreenLocation.IsZero())
-			{
-				continue;
-			}
-
-			DrawTag(Canvas, ScreenLocation, EntityId, Actor->GetName());
+			PlayerLocation = LocalPawn->GetActorLocation();
 		}
+
+		for (TPair<Worker_EntityId_Key, TWeakObjectPtr<AActor>>& EntityActorPair : EntityActorMapping)
+		{
+			const TWeakObjectPtr<AActor> Actor = EntityActorPair.Value;
+			const Worker_EntityId EntityId = EntityActorPair.Key;
+
+			if (Actor != nullptr)
+			{
+				FVector ActorLocation = Actor->GetActorLocation();
+
+				if (ActorLocation.IsZero())
+				{
+					continue;
+				}
+
+				if (FVector::Dist(PlayerLocation, ActorLocation) > MaxRange)
+				{
+					continue;
+				}
+
+				FVector2D ScreenLocation = FVector2D::ZeroVector;
+				if (LocalPlayerController.IsValid())
+				{
+					SCOPE_CYCLE_COUNTER(STAT_Projection);
+					UGameplayStatics::ProjectWorldToScreen(LocalPlayerController.Get(), ActorLocation + WorldSpaceActorTagOffset,
+														   ScreenLocation, false);
+				}
+
+				if (ScreenLocation.IsZero())
+				{
+					continue;
+				}
+
+				DrawTag(Canvas, ScreenLocation, EntityId, Actor->GetName(), true /*bCentre*/);
+			}
+		}
+	}
+}
+
+void GetReplicatedActorsInHierarchy(const AActor* Actor, TArray<const AActor*>& HierarchyActors)
+{
+	if (Actor->GetIsReplicated() && !HierarchyActors.Contains(Actor))
+	{
+		HierarchyActors.Add(Actor);
+	}
+	for (const AActor* Child : Actor->Children)
+	{
+		GetReplicatedActorsInHierarchy(Child, HierarchyActors);
 	}
 }
 
@@ -506,18 +677,20 @@ void ASpatialDebugger::DrawDebugLocalPlayer(UCanvas* Canvas)
 		return;
 	}
 
-	const TArray<TWeakObjectPtr<AActor>> LocalPlayerActors = { LocalPawn, LocalPlayerController, LocalPlayerState };
+	TArray<const AActor*> ActorsToDisplay = { LocalPlayerState.Get(), LocalPlayerController.Get(), LocalPawn.Get() };
+
+	if (bShowPlayerHierarchy)
+	{
+		GetReplicatedActorsInHierarchy(LocalPlayerController.Get(), ActorsToDisplay);
+	}
 
 	FVector2D ScreenLocation(PlayerPanelStartX, PlayerPanelStartY);
 
-	for (int32 i = 0; i < LocalPlayerActors.Num(); ++i)
+	for (int32 i = 0; i < ActorsToDisplay.Num(); ++i)
 	{
-		if (LocalPlayerActors[i].IsValid())
-		{
-			const Worker_EntityId EntityId = NetDriver->PackageMap->GetEntityIdFromObject(LocalPlayerActors[i].Get());
-			DrawTag(Canvas, ScreenLocation, EntityId, LocalPlayerActors[i]->GetName());
-			ScreenLocation.Y -= PLAYER_TAG_VERTICAL_OFFSET;
-		}
+		const Worker_EntityId EntityId = NetDriver->PackageMap->GetEntityIdFromObject(ActorsToDisplay[i]);
+		DrawTag(Canvas, ScreenLocation, EntityId, ActorsToDisplay[i]->GetName(), false /*bCentre*/);
+		ScreenLocation.Y += PLAYER_TAG_VERTICAL_OFFSET;
 	}
 }
 
@@ -540,6 +713,11 @@ void ASpatialDebugger::SpatialToggleDebugger()
 			CreateWorkerRegions();
 		}
 	}
+}
+
+bool ASpatialDebugger::IsEnabled()
+{
+	return DrawDebugDelegateHandle.IsValid();
 }
 
 #if WITH_EDITOR
@@ -581,7 +759,7 @@ void ASpatialDebugger::EditorInitialiseWorkerRegions()
 {
 	WorkerRegions.Empty();
 
-	const UWorld* World = GetWorld();
+	const UWorld* World = GEditor->GetEditorWorldContext().World();
 	check(World != nullptr);
 
 	const UAbstractSpatialMultiWorkerSettings* MultiWorkerSettings =
@@ -606,12 +784,28 @@ void ASpatialDebugger::EditorInitialiseWorkerRegions()
 			const PhysicalWorkerName WorkerName = PhysicalWorkerName::Printf(TEXT("WorkerRegion%d%d%d"), i, i, i);
 			WorkerRegionInfo.Color = GetColorForWorkerName(WorkerName);
 			WorkerRegionInfo.Extents = LBStrategyRegion.Value;
-
 			WorkerRegions[i] = WorkerRegionInfo;
 		}
 	}
 
 	// Needed to clean up LoadBalanceStrategy memory, otherwise it gets duplicated exponentially
 	GEngine->ForceGarbageCollection(true);
+}
+
+void ASpatialDebugger::PostEditChangeProperty(FPropertyChangedEvent& PropertyChangedEvent)
+{
+	Super::PostEditChangeProperty(PropertyChangedEvent);
+
+	if (PropertyChangedEvent.Property != nullptr)
+	{
+		const FName PropertyName(PropertyChangedEvent.Property->GetFName());
+		if (PropertyName == GET_MEMBER_NAME_CHECKED(ASpatialDebugger, WorkerRegionHeight)
+			|| PropertyName == GET_MEMBER_NAME_CHECKED(ASpatialDebugger, WorkerRegionVerticalScale)
+			|| PropertyName == GET_MEMBER_NAME_CHECKED(ASpatialDebugger, WorkerRegionOpacity))
+
+		{
+			EditorRefreshWorkerRegions();
+		}
+	}
 }
 #endif // WITH_EDITOR