+++ SpatialGDK/Source/SpatialGDK/Public/Interop/SpatialOutputDevice.h
@@ -17,13 +17,15 @@ public:
 
 	void AddRedirectCategory(const FName& Category);
 	void RemoveRedirectCategory(const FName& Category);
-	void SetVerbosityFilterLevel(ELogVerbosity::Type Verbosity);
+	void SetVerbosityLocalFilterLevel(ELogVerbosity::Type Verbosity);
+	void SetVerbosityCloudFilterLevel(ELogVerbosity::Type Verbosity);
 	void Serialize(const TCHAR* InData, ELogVerbosity::Type Verbosity, const FName& Category) override;
 
 	static Worker_LogLevel ConvertLogLevelToSpatial(ELogVerbosity::Type Verbosity);
 
 protected:
-	ELogVerbosity::Type FilterLevel;
+	ELogVerbosity::Type LocalFilterLevel;
+	ELogVerbosity::Type CloudFilterLevel;
 	TSet<FName> CategoriesToRedirect;
 	USpatialWorkerConnection* Connection;
 	FName LoggerName;