+++ SpatialGDK/Source/SpatialGDK/Public/LoadBalancing/LayeredLBStrategy.h
@@ -46,6 +46,8 @@ public:
 
 	virtual SpatialGDK::QueryConstraint GetWorkerInterestQueryConstraint() const override;
 
+	virtual bool RequiresHandoverData() const override;
+
 	virtual FVector GetWorkerEntityPosition() const override;
 
 	virtual uint32 GetMinimumRequiredWorkers() const override;