+++ SpatialGDK/Source/SpatialGDKTests/SpatialGDK/LoadBalancing/LayeredLBStrategy/LayeredLBStrategyTest.cpp
@@ -4,6 +4,7 @@
 #include "LoadBalancing/GridBasedLBStrategy.h"
 #include "LoadBalancing/LayeredLBStrategy.h"
 #include "LoadBalancing/SpatialMultiWorkerSettings.h"
+#include "SpatialConstants.h"
 #include "TestLayeredLBStrategy.h"
 #include "Utils/LayerInfo.h"
 
@@ -204,7 +205,7 @@ bool FCleanup::Update()
 
 LAYEREDLBSTRATEGY_TEST(GIVEN_strat_is_not_ready_WHEN_local_virtual_worker_id_is_set_THEN_is_ready)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -221,7 +222,7 @@ LAYEREDLBSTRATEGY_TEST(GIVEN_strat_is_not_ready_WHEN_local_virtual_worker_id_is_
 LAYEREDLBSTRATEGY_TEST(
 	GIVEN_layered_strat_of_two_by_four_grid_strat_singleton_strat_and_default_strat_WHEN_get_minimum_required_workers_called_THEN_ten_returned)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -239,7 +240,7 @@ LAYEREDLBSTRATEGY_TEST(
 LAYEREDLBSTRATEGY_TEST(
 	Given_layered_strat_of_2_single_cell_strats_and_default_strat_WHEN_set_virtual_worker_ids_called_with_2_ids_THEN_error_is_logged)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -260,7 +261,7 @@ LAYEREDLBSTRATEGY_TEST(
 LAYEREDLBSTRATEGY_TEST(
 	Given_layered_strat_of_2_single_cell_grid_strats_and_default_strat_WHEN_set_virtual_worker_ids_called_with_3_ids_THEN_no_error_is_logged)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -278,7 +279,7 @@ LAYEREDLBSTRATEGY_TEST(
 LAYEREDLBSTRATEGY_TEST(
 	Given_layered_strat_of_2_single_cell_grid_strats_and_no_handover_default_strat_WHEN_set_virtual_worker_ids_called_with_3_ids_THEN_no_error_is_logged_and_handover_is_disabled)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -308,9 +309,9 @@ LAYEREDLBSTRATEGY_TEST(Given_layered_strat_of_default_strat_WHEN_requires_handov
 	return true;
 }
 
-LAYEREDLBSTRATEGY_TEST(Given_layered_strat_of_single_cell_grid_strat_and_default_strat_WHEN_requires_handover_called_THEN_returns_true)
+LAYEREDLBSTRATEGY_TEST(Given_layered_strat_of_single_cell_grid_strat_and_default_strat_WHEN_requires_handover_called_THEN_returns_false)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -320,13 +321,34 @@ LAYEREDLBSTRATEGY_TEST(Given_layered_strat_of_single_cell_grid_strat_and_default
 	ADD_LATENT_AUTOMATION_COMMAND(FWaitForWorld(Data));
 	ADD_LATENT_AUTOMATION_COMMAND(FCreateStrategy(Data, Config, {}));
 	ADD_LATENT_AUTOMATION_COMMAND(FSetLocalVirtualWorker(Data, 1));
+	ADD_LATENT_AUTOMATION_COMMAND(FCheckRequiresHandover(Data, this, false));
+
+	return true;
+}
+
+LAYEREDLBSTRATEGY_TEST(Given_layered_strat_of_multiple_single_cell_grid_strategies_WHEN_requires_handover_called_THEN_returns_false)
+{
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
+
+	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
+
+	USpatialMultiWorkerSettings* MultiWorkerSettings = NewObject<USpatialMultiWorkerSettings>();
+	MultiWorkerSettings->WorkerLayers.Add(
+		FLayerInfo{ TEXT("LayerOne"), { ALayer1Pawn::StaticClass() }, USingleWorkerStrategy::StaticClass() });
+	MultiWorkerSettings->WorkerLayers.Add(
+		FLayerInfo{ TEXT("LayerTwo"), { ALayer2Pawn::StaticClass() }, USingleWorkerStrategy::StaticClass() });
+
+	ADD_LATENT_AUTOMATION_COMMAND(FWaitForWorld(Data));
+	ADD_LATENT_AUTOMATION_COMMAND(FCreateStrategy(Data, MultiWorkerSettings, {}));
+	ADD_LATENT_AUTOMATION_COMMAND(FSetLocalVirtualWorker(Data, 1));
+	ADD_LATENT_AUTOMATION_COMMAND(FCheckRequiresHandover(Data, this, false));
 
 	return true;
 }
 
 LAYEREDLBSTRATEGY_TEST(Given_layered_strat_of_default_strat_WHEN_who_should_have_auth_called_THEN_return_1)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -343,7 +365,7 @@ LAYEREDLBSTRATEGY_TEST(Given_layered_strat_of_default_strat_WHEN_who_should_have
 
 LAYEREDLBSTRATEGY_TEST(Given_layered_strat_WHEN_set_local_worker_called_twice_THEN_an_error_is_logged)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -361,7 +383,7 @@ LAYEREDLBSTRATEGY_TEST(Given_layered_strat_WHEN_set_local_worker_called_twice_TH
 
 LAYEREDLBSTRATEGY_TEST(Given_two_actors_of_same_type_at_same_position_WHEN_who_should_have_auth_called_THEN_return_same_for_both)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);
@@ -387,7 +409,7 @@ LAYEREDLBSTRATEGY_TEST(Given_two_actors_of_same_type_at_same_position_WHEN_who_s
 LAYEREDLBSTRATEGY_TEST(
 	GIVEN_two_actors_of_different_types_and_same_positions_managed_by_different_layers_WHEN_who_has_auth_called_THEN_return_different_values)
 {
-	AutomationOpenMap("/Engine/Maps/Entry");
+	AutomationOpenMap(SpatialConstants::EMPTY_TEST_MAP_PATH);
 
 	TSharedPtr<TestData> Data = TSharedPtr<TestData>(new TestData);
 	TSharedPtr<TestConfiguration> Config = TSharedPtr<TestConfiguration>(new TestConfiguration);